feat: remove cholesterol
