+++ src/pages/Defi/components/OpportunityCard.tsx
@@ -21,7 +21,6 @@ import {
   osmosisChainId,
 } from '@shapeshiftoss/caip'
 import { PairIcons } from 'features/defi/components/PairIcons/PairIcons'
-import type { EarnOpportunityType } from 'features/defi/helpers/normalizeOpportunity'
 import qs from 'qs'
 import { useCallback } from 'react'
 import { useHistory } from 'react-router'
@@ -33,6 +32,7 @@ import { RawText, Text } from 'components/Text'
 import { WalletActions } from 'context/WalletProvider/actions'
 import { useWallet } from 'hooks/useWallet/useWallet'
 import type { AssetsById } from 'state/slices/assetsSlice/assetsSlice'
+import type { EarnOpportunityType } from 'state/slices/opportunitiesSlice/types'
 import { selectAssetById, selectAssets } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
