+++ src/state/slices/opportunitiesSlice/types.ts
@@ -1,6 +1,5 @@
-import type { AccountId, AssetId } from '@shapeshiftoss/caip'
+import type { AccountId, AssetId, ChainId } from '@shapeshiftoss/caip'
 import type { DefiProvider, DefiType } from 'features/defi/contexts/DefiManagerProvider/DefiCommon'
-import type { EarnOpportunityType } from 'features/defi/helpers/normalizeOpportunity'
 import type { PartialRecord } from 'lib/utils'
 import type { Nominal } from 'types/common'
 
@@ -132,6 +131,32 @@ export type GetOpportunityUserStakingDataOutput = {
 }
 
 export type GetOpportunityIdsOutput = OpportunityId[]
+
+// TODO: This is not FDA-approved and should stop being consumed to make things a lot tidier without the added cholesterol
+export type EarnOpportunityType = {
+  type?: string
+  provider: string
+  version?: string
+  contractAddress?: string
+  rewardAddress: string
+  apy?: number | string
+  tvl: string
+  underlyingAssetId?: AssetId
+  assetId: AssetId
+  id: OpportunityId
+  fiatAmount: string
+  /** @deprecated use cryptoAmountBaseUnit instead and derive precision amount from it*/
+  cryptoAmountPrecision: string
+  cryptoAmountBaseUnit: string
+  expired?: boolean
+  chainId: ChainId
+  showAssetSymbol?: boolean
+  isLoaded: boolean
+  icons?: string[]
+  // overrides any name down the road
+  opportunityName?: string
+  highestBalanceAccountAddress?: string // FOX/ETH specific, let's change it to accountId across the line if we need it for other opportunities
+}
 
 export type StakingEarnOpportunityType = OpportunityMetadata & {
   stakedAmountCryptoBaseUnit?: string
